feature: merge support service

+++ microservice/microservice-support-parent/microservice-support-client/src/main/java/com/vinesmario/microservice/client/storage/dto/StorageImageDto.java
@@ -51,7 +51,7 @@ public class StorageImageDto extends BaseDto<Long> {
     /**
      * 图片大小，单位B
      */
-    private Integer imageSize;
+    private Long imageSize;
     /**
      * 缩略图名称
      */
