+++ microservice/microservice-support-parent/microservice-support-client/src/main/java/com/vinesmario/microservice/client/storage/dto/condition/StorageImageConditionDto.java
@@ -65,7 +65,7 @@ public class StorageImageConditionDto implements ConditionDto {
     /**
      * 图片大小，单位B
      */
-    private Integer imageSize;
+    private Long imageSize;
     /**
      * 缩略图名称
      */
