+++ microservice/microservice-support-parent/microservice-support-server/src/main/java/com/vinesmario/microservice/server/storage/entity/StorageImage.java
@@ -50,7 +50,7 @@ public class StorageImage extends BaseEntity<Long> {
     /**
      * 图片大小，单位B
      */
-    private Integer imageSize;
+    private Long imageSize;
     /**
      * 缩略图名称
      */
